[refactor] DDD 설계 원칙에 맞춰 디렉토리 구조 변경
