+++ notification/src/main/java/com/hubEleven/notification/ai/domain/entity/RequestStatus.java
@@ -1,4 +1,4 @@
-package com.hubEleven.notification.ai.domain.model;
+package com.hubEleven.notification.ai.domain.entity;
 
 import lombok.Getter;
